Add option to set render condition callback on action (#597)
