+++ src/Column/Action.php
@@ -20,6 +20,7 @@ class Action extends Column
 	use Traits\TButtonTryAddIcon;
 	use Traits\TButtonText;
 	use Traits\TLink;
+	use Traits\TRenderCondition;
 
 	/**
 	 * @var string
@@ -109,6 +110,10 @@ class Action extends Column
 	 */
 	public function render(Row $row)
 	{
+		if (!$this->shouldBeRendered($row)) {
+			return null;
+		}
+
 		try {
 			// Renderer function may be used
 			return $this->useRenderer($row);
