+++ src/Column/ItemDetail.php
@@ -23,6 +23,7 @@ class ItemDetail
 	use Traits\TButtonClass;
 	use Traits\TButtonTitle;
 	use Traits\TButtonText;
+	use Traits\TRenderCondition;
 
 	/**
 	 * (renderer | template | block)
@@ -59,11 +60,6 @@ class ItemDetail
 	 * @var array
 	 */
 	protected $template_parameters = [];
-
-	/**
-	 * @var callable
-	 */
-	protected $render_condition_callback;
 
 
 	/**
@@ -238,29 +234,5 @@ class ItemDetail
 	public function getTemplateVariables()
 	{
 		return $this->template_parameters;
-	}
-
-
-	/**
-	 * @param callable $condition
-	 * @return static
-	 */
-	public function setRenderCondition(callable $condition)
-	{
-		$this->render_condition_callback = $condition;
-
-		return $this;
-	}
-
-
-	/**
-	 * @param Row $row
-	 * @return bool
-	 */
-	public function shouldBeRendered(Row $row)
-	{
-		$condition = $this->render_condition_callback;
-
-		return $condition ? $condition($row->getItem()) : true;
 	}
 }
